Hide the save tab unless the user is logged in
We use mustache to include the save tab in the HTML only when a user is
logged in. Users can still access the save page by using the URI, in
which case they get the default message about logging in, but you still
will not see the save tab. This could have been done using
.hide()/.show() but the logic for this is more complex; the current code
is based on the observation that whenever log in or log out is selected
the page is reloaded (so the template approach looks feasible).  I am
not convinced this is a sensible change, hence the use of the
hidesavetab branch.