Use a simple pubsub implementation to create the saved-search page via Ajax/client rather than Mustache/server
I used a simple pub-sub implementation from
http://addyosmani.com/largescalejavascript/ to send user/login and
user/logout messages. These messages are created in semflow.core by
querying the server to find out if a user is logged in or not. The code
in saved.js responds to the messages to determine what to display and
creates the tables when needed. This requires returning more information
from the server (using the new /savedsearches2 and /savedpubs2
endpoints).